Bill S504A version A - Gustavo Rivera
BILL NUMBER: S504A SPONSOR: RIVERA TITLE OF BILL: An act to amend the insurance law, in relation to prohibiting cost shar- ing for insulin SUMMARY OF PROVISIONS: Sections 1 through 3 of the bill amend various sections of insurance law which were amended as part DDD of Chapter 56 of the laws of 2020; the adopted budget, to eliminate any deductible, copayment, coinsurance or any other cost sharing requirement for insulin drugs. Section 4 of the bill provides for an effective date.