Bill S8399 version original - Jabari Brisport
BILL NUMBER: S8399 SPONSOR: BRISPORT TITLE OF BILL: An act to amend the real property law, in relation to prohibiting the recording of conveyances suspected to be fraudulent SUMMARY OF PROVISIONS: Section one amends Section 291 of the real property law, to provide that a county registrar shall not record a conveyance if the county registrar has reason to believe that the conveyance is false or fraudulent in any manner. Section two sets forth the effective date.