Bill S8390 version original - Jeremy Cooney
BILL NUMBER: S8390 SPONSOR: COONEY TITLE OF BILL: An act to amend the criminal procedure law and the civil practice law and rules, in relation to the admissibility of evidence created or proc- essed by artificial intelligence PURPOSE OR GENERAL IDEA OF BILL: Relates to the admissibility of evidence created or processed by artifi- cial intelligence. SUMMARY OF SPECIFIC PROVISIONS: 60.