Bill S8401 version original - Pamela Helming
BILL NUMBER: S8401 SPONSOR: HELMING TITLE OF BILL: An act to amend the public health law, in relation to the prohibition of the sale of kratom to individuals under the age of twenty-one PURPOSE OR GENERAL IDEA OF BILL: To prohibit the sale of kratom to individuals who are under the age of twenty-one. SUMMARY OF SPECIFIC PROVISIONS: Sections 1 and 2: The public health law is amended to provide for the regulation of Kratom and defines what Kratom is.