Bill S8423 version original - James Skoufis
BILL NUMBER: S8423 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the village law, in relation to removing the ability of a board of trustees and mayor to act as the board of assessors PURPOSE: To remove the ability of a board of trustees and mayor to act as the board of assessors. SUMMARY OF PROVISIONS: Section 1: This section allows the board of trustees to consolidate the offices of clerk, treasurer, and assessor, or any of these offices, within a village.