Bill S8421 version original - Neil D. Breslin
BILL NUMBER: S8421 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the civil service law, in relation to increasing the age of eligibility for provisional or permanent appointment of certain police officers from thirty-five to forty PURPOSE OR GENERAL IDEA OF BILL: To increase the maximum age that a police officer can serve. SUMMARY OF PROVISIONS: Section one amends section 58 of the civil service law to raise the maximum age to allow police officers to serve until they are forty years old.