Bill S8431 version original - James Skoufis
BILL NUMBER: S8431 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the environmental conservation law, in relation to requiring nuisance wildlife control operators to disclose that euthana- sia may be used to resolve a wildlife issue PURPOSE: To require that before a nuisance wildlife control operator takes any control measures, they must provide the client with a written statement. This statement must indicate that euthanasia may be used as a method to resolve the wildlife issue.