Bill S8433 version original - Jessica Scarcella-Spanton
BILL NUMBER: S8433 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the penal law and the correction law, in relation to establishing the crime of harassment of an employee by an incarcerated individual PURPOSE: Establishes the crime of harassment of an employee by an incarcerated individual. SUMMARY OF PROVISIONS: Section 1: Amends the penal law by adding a new section 240.33 creating the crime of Harassment of an employee by an incarcerated individual.