Bill S8443 version original - Anthony H. Palumbo
BILL NUMBER: S8443 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the alcoholic beverage control law, in relation to exempting certain parcels of land from licensing restrictions for manufacturers and wholesalers of alcoholic beverages who sell at retail for on-premises consumption PURPOSE: To amend Chapter 328 of the Laws of 2022 by providing the metes and bounds of the new location of Talkhouse Encore, in the town of East Hampton. SUMMARY OF PROVISIONS: Section 1.