Bill S8446B version original - Neil D. Breslin
BILL NUMBER: S8446 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the election law, in relation to certain party nomi- nations for candidates for county office PURPOSE: To provide that party designation of a candidate for nomination for any office be filled by the voters of the entire county shall be made by the county committee and eliminate unnecessary petitioning for designated candidates. SUMMARY OF PROVISIONS: Section 1.