Bill S7754A version A - Neil D. Breslin
BILL NUMBER: S7754A SPONSOR: BRESLIN TITLE OF BILL: An act to amend the penal law, in relation to the crime of staging a motor vehicle accident in the second degree PURPOSE: To impose criminal penalties on those who recruit, hire, or otherwise direct others to engage in staged accidents. SUMMARY OF PROVISIONS: Section 1 amends section 176,.75 of the penal law. Section 2 provides.an effective date.