Bill S8439 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S8439 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the racing, pari-mutuel wagering and breeding law, in relation to authorizing one percent of mobile sports tax revenue be used for problem gambling PURPOSE OR GENERAL IDEA OF BILL: To authorize one percent of mobile sports tax revenue to be used for addiction services. SUMMARY OF PROVISIONS: Section 1 of this bill amends subdivision 8 of section 1367 of the racing, pari-mutuel wagering and breeding law.