Bill S8437A version A - Iwen Chu
BILL NUMBER: S8437A SPONSOR: CHU TITLE OF BILL: An act to establish a commission for the educational advancement of multilingual learners in New York; and providing for the repeal of such provisions upon expiration thereof PURPOSE: The purpose of this legislation is to establish a commission for the educational advancement of multilingual learners. SUMMARY OF PROVISIONS: Section One establishes the commission for the educational advancement of multilingual learners.