Bill S8451 version original - Jessica Ramos
BILL NUMBER: S8451 SPONSOR: RAMOS TITLE OF BILL: An act to amend the labor law and the alcoholic beverage control law, in relation to suspending certain liquor licenses of employers who are in violation of certain provisions of the labor law PURPOSE: Amends the labor and the alcoholic beverages control law to suspend certain liquor licenses of employers violating certain provisions of the labor law. SUMMARY OF PROVISIONS: Section 1.