Bill S8373A version A - Gustavo Rivera
BILL NUMBER: S8373A SPONSOR: RIVERA TITLE OF BILL: An act to amend the general business law and the public health law, in relation to the definition of medical debt PURPOSE OR GENERAL OF BILL: Relates to the definition of medical debt. SUMMARY OF SPECIFIC PROVISIONS: Section 1 would amend subdivision (v) of section 380-a of the general business law, as it relates to the Fair Credit Reporting Act; Defi- nitions to define the term "medical debt.