Bill S8452A version original - Jessica Ramos
BILL NUMBER: S8452 SPONSOR: RAMOS TITLE OF BILL: An act to amend the labor law and the workers' compensation law, in relation to the issuing of stop-work orders PURPOSE: Amends the labor law and the worker's compensation law, in relation to the issuing of stop-work orders. SUMMARY OF PROVISIONS: Section 1. This act shall be known and cited as the "stop workplace abuse today (SWAT) act". Section 2. Amends Section 21 of the labor law by adding new subdivision 17 to read.