Bill S7543B version A - Kristen Gonzalez
BILL NUMBER: S7543A SPONSOR: GONZALEZ PURPOSE: This bill would require the disclosure of automated decision-making systems already in use, prohibit the unauthorized use of automated deci- sion-making systems by state agencies or entities on behalf of an agency and, when authorized, require the publication of an impact assessment. SUMMARY OF PROVISIONS: Section 1. Provides the title, which is the "legislative oversight of automated decision-making in government act (LOADinG Act)." Section 2.