Bill S8453 version original - Jessica Ramos
BILL NUMBER: S8453 SPONSOR: RAMOS TITLE OF BILL: An act to amend the labor law and the tax law, in relation to suspending certain certificates of authority of employers who are in violation of certain provisions of the labor law PURPOSE: Amends the labor law and the tax law, suspending certificates of author- ity of employers who violate certain labor laws. SUMMARY OF PROVISIONS: Section 1.