Bill S8464 version original - James Skoufis
BILL NUMBER: S8464 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the election law, in relation to moving the date of the meeting of the electors and the method for the transmission of the certificates of vote PURPOSE: To amend the date of the meeting of the electoral college. SUMMARY OF PROVISIONS: Section 1: Amends Section 12-104 of the election law, as amended by chapter 234 of the laws of 1976.