Bill S5107C version B - Brad Hoylman-Sigal
BILL NUMBER: S5107B SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the family court act, the domestic relations law and the general business law, in relation to surrogacy programs and agreements SUMMARY OF PROVISIONS: Sections 1 through 20 make amendments to Article 5-C of the Family Court Act. Section 21 makes an amendment to section 123(2)(a) of the Domestic Relations Law. Sections 22 through 25 make amendments to Article 44 of the General Business Law.