Bill S8494 version original - Cordell Cleare
BILL NUMBER: S8494 SPONSOR: CLEARE TITLE OF BILL: An act in relation to creating a public benefit corporation for the construction, acquisition, and rehabilitation of permanently affordable housing; and making an appropriation therefor PURPOSE OR GENERAL IDEA OF BILL: To create a new state authority with the mandate to build, acquire, rehabilitate, and preserve permanently affordable housing statewide. SUMMARY OF SPECIFIC PROVISIONS: Section one contains the short title.