Bill S2898A version A - Zellnor Myrie
BILL NUMBER: S2898A SPONSOR: MYRIE TITLE OF BILL: An act to amend the public health law, in relation to the maternal health crisis and establishing two centers of maternal and childcare services; and to repeal title 3 of article 25 of such law relating to the control of midwifery PURPOSE: To establish a plan to address the severe maternal morbidity crisis in parts of the state. SUMMARY OF PROVISIONS: Section 1 provides the title of the act as the "Chisholm Chance Act".