Bill S8512 version original - Kevin Thomas
BILL NUMBER: S8512 SPONSOR: THOMAS TITLE OF BILL: An act to amend the real property tax law, in relation to tax lien fore- closure PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to implement procedures for returning surplus funds to homeowners in response to Tyler v. Hennepin County and to provide homeowners at risk of tax lien foreclosures the same protections that are afforded to borrowers in residential mortgage fore- closure proceedings.