Bill S5485B version B - Nathalia Fernandez
BILL NUMBER: S5485B SPONSOR: FERNANDEZ PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to replace the word "addict" and "addicts" with modern and humane language. SUMMARY OF SPECIFIC PROVISIONS: Sections one through four and six through twelve replace the word "addict" and "narcotic addict" with "person with substance use disorder" and variations thereof.