Bill S8534A version original - Brad Hoylman-Sigal
BILL NUMBER: S8534 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to prohibiting the use of virtual reality devices while operating a motor vehicle PURPOSE OF BILL: To prohibit the use of virtual reality devices while operating a motor vehicle. SUMMARY OF SPECIFIC PROVISIONS: Section one amends Subdivision 1 and paragraphs (a) and (b) of subdivi- sion 2 of section 1225-d of the vehicle and traffic law.