Bill S8539 version original - Jessica Scarcella-Spanton
BILL NUMBER: S8539 Retired 3/27/24 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the penal law, in relation to categorizing the assault of certain employees of the New York city department of buildings or the New York city department of housing preservation and development as a class D felony SUMMARY OF PROVISIONS: Section 1. Section 2.