Bill S6009A version A - Kevin S. Parker
BILL NUMBER: S6009A SPONSOR: PARKER TITLE OF BILL: An act to amend the mental hygiene law, in relation to creating a substance use disorder treatment bill of rights PURPOSE OR GENERAL IDEA OF BILL: To ensure that patients of substance use disorder treatment facilities are accorded appropriate rights. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Subdivision (c) of section 22.03 if the mental hygiene law, as added by chapter 558 of the laws of 1999, is amended and a new subdi- vision (d) is added.