Bill S8541 version original - Kevin S. Parker
BILL NUMBER: S8541 SPONSOR: PARKER TITLE OF BILL: An act to amend the public authorities law, in relation to authorizing the New York state energy research and development authority to adminis- ter a program to provide funds related to enabling clean wind projects PURPOSE OR GENERAL IDEA OF BILL: AN ACT that will enable New York State Energy & Research to aid in the creation of an incentivized program to provide funds related to enabling clean energy projects. SUMMARY OF PROVISIONS: 1.