Bill S1050B version A - Jeremy Cooney
BILL NUMBER: S1050A SPONSOR: COONEY PURPOSE: To remove the current limitation on the use of pharmacy technicians in pharmacies across New York, in order to support independent pharmacies. SUMMARY OF PROVISIONS: Section 1. Amends subdivision 1 of section 6841 of the education law, as added by chapter 414 of the laws of 2019. Section 2. Sets the effective date. JUSTIFICATION: Pharmacy technicians are essential to the effective and safe adminis- tration of pharmacies throughout the State.