Bill S6956A version A - Zellnor Myrie
BILL NUMBER: S6956A SPONSOR: MYRIE TITLE OF BILL: An act in relation to requiring the commissioner of health to develop a sustainability plan for the state university of New York downstate medical center PURPOSE: To require the commissioner of the Department of Health to draft and submit a sustainability plan for the SUNY Downstate Medical Center. SUMMARY OF PROVISIONS: Section 1 provides the legislative findings and intent.