Bill S8558 version original - Cordell Cleare
BILL NUMBER: S8558 SPONSOR: CLEARE TITLE OF BILL: An act to amend the executive law, in relation to establishing a coordi- nator for asylum seeker services PURPOSE OR GENERAL IDEA OF BILL: This bill will establish a Coordinator of Asylum Seeker Services to implement a statewide plan for the intake and resettlement of asylum seekers. SUMMARY OF PROVISIONS: Section 1 amends the executive law by adding a new section4-c to estab- lish a coordinator for asylum seeker services, and defines terms.