Bill S8559 version original - Leroy Comrie
BILL NUMBER: S8559 SPONSOR: COMRIE TITLE OF BILL: An act to amend the tax law, in relation to the taxation of vapor products PURPOSE: To strengthen enforcement of the ban on flavored vape products. SUMMARY OF PROVISIONS: Section 1. Amends section 1180 of the tax law by adding a new subdivi- sion (c) Section 2. Amends section 1181 of the tax law, as chapter 92 of the laws of 2021. Section 3. Amends the tax law by adding a new section 1183-a. Section 4.