Bill S8555 version original - Jessica Ramos
BILL NUMBER: S8555 SPONSOR: RAMOS TITLE OF BILL: An act to amend the labor law, in relation to employers' obligations to report wages for the purposes of unemployment benefits, the payment of unemployment benefits, and employer penalties for non-compliance PURPOSE: The purpose of this bill is to ensure that all employers that have been ruled covered by New York's unemployment insurance laws report their employees' wages and contribute to the UI Trust Fund. SUMMARY OF PROVISIONS: Section One.