Bill S8564 version original - Monica Martinez
BILL NUMBER: S8564 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the labor law, in relation to nursing employees PURPOSE: The purpose of this legislation is to expand the rights afforded to nursing employees to include the provision of a sink or individual basin used solely for the washing of breast pump attachments upon request by the employee. SUMMARY OF PROVISIONS: Section 1.