Bill S8570 version original - Mario Mattera
BILL NUMBER: S8570 SPONSOR: MATTERA TITLE OF BILL: An act directing the New York state energy research and development authority to conduct a feasibility study on sewage thermal energy PURPOSE: To have the New York state energy research and development authority to conduct a feasibility study and prepare a report on the benefits and implementation of sewage thermal energy, to study the potential impact and cost. SUMMARY OF PROVISIONS: Section 1.