Bill S8568 version original - Mario Mattera
BILL NUMBER: S8568 SPONSOR: MATTERA TITLE OF BILL: An act directing the New York state energy research and development authority to conduct a feasibility study on the use of small modular reactors PURPOSE: To have the New York state energy research and development authority to conduct a feasibility study and prepare a report on the benefits and implementation of use of small modular reactors, to study the potential impact and cost. SUMMARY OF PROVISIONS: Section 1.