Bill S7998A version A - Leroy Comrie
BILL NUMBER: S7998A SPONSOR: COMRIE TITLE OF BILL: An act to amend the cannabis law and the penal law, in relation to forfeiture of property involved in the unlawful sale of cannabis PURPOSE OR GENERAL IDEA OF BILL: To authorize all municipalities to stop operations of unlicensed canna- bis retailers and seize merchandise. SUMMARY OF PROVISIONS: Section 1 is the short title.