Bill S8578 version original - Pamela Helming
BILL NUMBER: S8578 SPONSOR: HELMING TITLE OF BILL: An act to amend the tax law, in relation to establishing a housing infrastructure tax credit PURPOSE OR GENERAL IDEA OF BILL: To provide a tax credit of up to ten percent of the cost for infrastruc- ture projects related to the construction of new homes or multiple dwellings. SUMMARY OF SPECIFIC PROVISIONS: Sections 1: The tax law is amended to add a new section 49 establishing the housing infrastructure tax credit.