Bill S8573 version original - Lea Webb
BILL NUMBER: S8573 SPONSOR: WEBB TITLE OF BILL: An act to amend the general business law, in relation to the practice of natural hair care and braiding PURPOSE: Relates to the practice of natural hair care and braiding. SUMMARY OF PROVISIONS: Section 1. The heading of article 27 of the general business law, is updated by removing "styling" and inserting "care and braiding" to read "Licensing of Nail Specialty, Natural Hair and Braiding, Esthetics and Cosmetology." Section 2.