Bill S8566 version original - Mario Mattera
BILL NUMBER: S8566 SPONSOR: MATTERA TITLE OF BILL: An act directing the New York state energy research and development authority to conduct a feasibility study on geothermal energy PURPOSE: To have the New York state energy research and development authority to conduct a feasibility study and prepare a report on the benefits and implementation of use of geothermal energy, to study the potential impact and cost. SUMMARY OF PROVISIONS: Section 1.