Bill S8478A version A - Iwen Chu
BILL NUMBER: S8478A SPONSOR: CHU TITLE OF BILL: An act to amend the education law, in relation to the library materials factor PURPOSE: The purpose of this legislation is to adjust the per-pupil rate of school library material aid. SUMMARY OF PROVISIONS: Section 1 amends the education law by raising the per-pupil cost of school library materials aid for the 2025-2026 school year, and adjusts the continuing rate based on the Consumer Price Index. Section 2 provides the effective date.