Bill S8592 version original - Luis R. Sepúlveda
BILL NUMBER: S8592 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the state finance law, in relation to including the city of New York in the aid and incentives for municipalities program PURPOSE: To add New York City back into the aid and incentives for municipalities program SUMMARY OF PROVISIONS: Section 1. The opening paragraph j of subdivision 10 of section 54 of the state finance law, is amended to include the City of New York.