Bill S8601A version original - Monica Martinez
BILL NUMBER: S8601 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the labor law, in relation to prohibiting mandatory iris and retina scanning of employees PURPOSE: The purpose of this bill is to prevent employers from requiring employ- ees to undergo iris and/or retina scanning to sustain or gain employ- ment, with exemptions provided. SUMMARY OF PROVISIONS: Section 1. amends the labor law adding section 201-a-1 to prohibit iris and retina scanning of employees.