Bill S8616 version original - Kevin S. Parker
BILL NUMBER: S8616 SPONSOR: PARKER TITLE OF BILL: An act to amend the public authorities law, in relation to increasing development of clean energy distribution infrastructure PURPOSE OR GENERAL IDEA OF BILL: To increase the development of clean energy infrastructure across the state and improve transmission rates. SUMMARY OF PROVISIONS: Section 1. Article 8 of the public authorities' law is amended by 2 adding a new title 9-D Section 2. 1920. Definitions. As used in this title Section 3. 1921.