Bill S8609 version original - Mark Walczyk
BILL NUMBER: S8609 SPONSOR: WALCZYK TITLE OF BILL: An act to amend the penal law, in relation to allowing certain firearm licensees to open carry PURPOSE OR GENERAL IDEA OF BILL: This bill would amend the penal law to allow New Yorkers to open carry firearms. SUMMARY OF PROVISIONS: Section 1: Amends chapter 371 of 2022 to include open carry as an option for individuals who are filling out a firearm permit, in addition to concealed carry. Section 2: Sets the effective date.