Bill S8625 version original - Jack M. Martins
BILL NUMBER: S8625 SPONSOR: MARTINS TITLE OF BILL: An act to amend the penal law, in relation to enacting the "New York state antisemitism vandalism act" PURPOSE: To make it unlawful for individuals to knowingly destroy, damage, remove any banner, poster, flyer, or billboard which is in a public space and/or is pro-Israel. SUMMARY OF PROVISIONS: Section 1. establishes the act as the "New York State Antisemitism Vandalism Act." Section 2. amends the penal law by adding a new section 240.80.