Bill S8637 version original - Iwen Chu
BILL NUMBER: S8637 SPONSOR: CHU TITLE OF BILL: An act to amend the election law, in relation to authorizing absentee voting for religious observance PURPOSE: The purpose of this legislation is to allow for absentee voting in the event of a religious observance. SUMMARY OF PROVISIONS: Section 1 amends section 8-400 of the election law, adding new subsection (e) to allow for absentee voting due to a religious obser- vance. Section 2 provides the effective date.