Bill S8339A version A - Shelley Mayer
BILL NUMBER: S8339A SPONSOR: MAYER TITLE OF BILL: An act to amend the family court act, in relation to establishing stand- ards for public access to family court PURPOSE OF BILL: To clarify the standard to be used by a family court judge when deter- mining whether to close the courtroom to the public or exclude someone from the courtroom. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends the family court act by adding a new section 161-a. Section 2 amends § 341.