Bill S8653A version A - Michael Gianaris
BILL NUMBER: S8653A SPONSOR: GIANARIS TITLE OF BILL: An act to amend the state law, in relation to establishing congressional districts; and to repeal article 7 of such law relating thereto PURPOSE: The purpose of this bill is to establish 26 Congressional districts in and for the State of New York. SUMMARY OF PROVISIONS: This bill would provide metes & bounds descriptions of 26 Congressional districts.