Bill S8655 version original - Cordell Cleare
BILL NUMBER: S8655 SPONSOR: CLEARE TITLE OF BILL: An act to amend the social services law, in relation to the standards of monthly need for persons in receipt of public assistance PURPOSE GENERAL IDEA OF BILL: This legislation will increase the outdated personal and special needs allowances for certain individuals on public assistance. SUMMARY OF PROVISIONS: Section one of this bill amends section 131-a of the social services law. Section two provides the effective date.