Bill S8663 version original - Brad Hoylman-Sigal
BILL NUMBER: S8663 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the executive law, in relation to limiting recordkeeping and reporting duties of public notaries PURPOSE: Exempt notarial acts as they relate non-electronic notarization, from being recorded or maintained in a notarial record. SUMMARY OF PROVISIONS: Section 1: Creates a new Section 135-D of the Executive Law as it relates to notarial record keeping and reporting.