Bill S8688A version original - Jamaal Bailey
BILL NUMBER: S8688 SPONSOR: BAILEY TITLE OF BILL: An act to amend the criminal procedure law, in relation to facilitating appellate review of rulings that implicate issues of public concern PURPOSE: The purpose of the Act is to require the court to review a denial of a motion to suppress evidence on appeal. By ensuring that a court review such denials, the legislative intent of section 710.70 will be furthered, and defendants can properly challenge the validity of an order denying their motion.