Bill S8683A version original - Kevin S. Parker
BILL NUMBER: S8683 SPONSOR: PARKER TITLE OF BILL: An act to amend the cannabis law, in relation to employment at certain licensees for the purposes of parole conditions PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to allow persons currently required to be employed as a condition of parole, probation, or other state supervision to fulfill employment requirements through employment at a licensed cannabis business. SUMMARY OF PROVISIONS: Section 1.