Bill S8693 version original - Cordell Cleare
BILL NUMBER: S8693 SPONSOR: CLEARE TITLE OF BILL: An act to amend the public service law, in relation to requiring inve- stor-owned energy utilities to collect and publish data on disadvantaged communities in certain service areas PURPOSE OR GENERAL IDEA OF BILL: This bill requires investor-owned energy utilities to publish statistics about customers in disadvantaged communities on their website on an annual basis and file them with the commission. Violations will be penalized.