Bill S8700 version original - Cordell Cleare
BILL NUMBER: S8700 SPONSOR: CLEARE TITLE OF BILL: An act to amend the general business law, in relation to requiring that digital coupons are made available to consumers in paper form of identi- cal value PURPOSE: This legislation will require that digital coupons are made available to consumers in paper form of identical value. SUMMARY OF SPECIFIC PROVISIONS: Section.