Bill S1530A version A - Robert Ortt
BILL NUMBER: S1530A SPONSOR: ORTT TITLE OF BILL: An act to amend the executive law, in relation to the use of biodiesel and renewable diesel fuel in state vehicles PURPOSE: To mandate the state to use biodiesel in all state vehicles that are operated on diesel. SUMMARY OF SPECIFIC PROVISIONS: Section 1 - Establishes the legislative intent of the bill.