Bill S8725A version original - James Skoufis
BILL NUMBER: S8725 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the village of Washingtonville deficit financing act, in relation to the issuance of certain bonds PURPOSE: Extends the dates allowing the Village of Washingtonville in Orange County to issue serial bonds and extends the date up to which the Village can use the funds to address accumulated deficits in the general fund, water fund, sewer fund, and capital projects fund. SUMMARY OF PROVISIONS: Section 1.