Bill S8707 version original - Jamaal Bailey
BILL NUMBER: S8707 SPONSOR: BAILEY TITLE OF BILL: An act to amend the education law, in relation to permitting certain licensed athletic trainers to practice in New York state if they are licensed to practice in another state or territory PURPOSE: To allow licensed trainers to practice in the state of New York if they are licensed in another state. SUMMARY OF PROVISIONS: Section 1. Section 8356 of the education law is amended by adding a new subdivision 4. Section 2. Sets the effective date.