Bill S8713 version original - Jessica Scarcella-Spanton
BILL NUMBER: S8713 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the criminal procedure law, in relation to adolescent offenders authorized to be removed to family court PURPOSE OR GENERAL IDEA OF BILL: This bill would remove the requirement to have an adolescent "display" a loaded firearm for them to illegally possess it. SUMMARY OF PROVISIONS: Section 1. Amend subparagraph (ii) of paragraph (c) of subdivision 2 of section 722.23 of the criminal procedure law. Section 2.