Bill S8721 version original - Timothy M. Kennedy
BILL NUMBER: S8721 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the education law, in relation to declaring Pulaski Day a school holiday PURPOSE: This legislation would declare Pulaski Day a school holiday. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends subdivision 8 of section 3604 of the education law to declare October eleventh known as Pulaski Day, as a public school holi- day. Section 2 amends subdivision 2-a of section 3635 of the education law to make a conforming change.