Bill S8723 version original - Thomas F. O'Mara
BILL NUMBER: S8723 SPONSOR: O'MARA TITLE OF BILL: An act to amend the retirement and social security law, in relation to the retirement of traffic officers with the town of Elmira PURPOSE OR GENERAL IDEA OF BILL: To allow Elmira traffic officers to opt into the 20-year police and fire retirement system. SUMMARY OF PROVISIONS: Section 1. authorizes Elmira traffic officers to opt into the 20-year police and fire retirement system. Section 2. states the requirements for retirement. Section 3.