Bill S8726B version original - Robert Jackson
BILL NUMBER: S8726 SPONSOR: JACKSON TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to abandoned vehicles PURPOSE: Allow for abandoned vehicles to be removed off City of New York roads. CURRENT LAW: Section 1224(1)(a) of the Vehicle and Traffic Law ("VTL") currently allows for a motor vehicle to be deemed abandoned if it is left unat- tended with no number plates affixed for more than six hours on any highway or other public place.