Bill S8732 version original - Zellnor Myrie
BILL NUMBER: S8732 SPONSOR: MYRIE TITLE OF BILL: An act to amend the executive law, in relation to establishing the commission for the modernization and revitalization of downstate medical center PURPOSE: Establishes the commission for the modernization and revitalization of Downstate Medical Center. SUMMARY OF PROVISIONS: Section 1 amends the executive law by adding article 49-c which estab- lishes the commission for the modernization and revitalization of Down- state Medical Center.