Bill S4890A version A - Gustavo Rivera
BILL NUMBER: S4890A SPONSOR: RIVERA TITLE OF BILL: An act to amend the public health law, in relation to requiring chain restaurants to label menu items that have a high content of sodium PURPOSE: To require chain restaurants to label menu items that have a high content of sodium. SUMMARY OF SPECIFIC PROVTSIONS: Section 1 amends the public health law by adding a new section 1357 as it relates to chain restaurants and the sodium content of menu items.