Bill S8738 version original - Peter Oberacker
BILL NUMBER: S8738 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the executive law, in relation to establishing an adopt-a-rural cemetery program PURPOSE: Requires the Secretary of State to develop and implement an "Adopt-A- Rural Cemetery" program to improve, repair and maintain cemeteries in rural communities. SUMMARY OF PROVISIONS: Section 1.