Bill S8734A version original - Cordell Cleare
BILL NUMBER: S8734 SPONSOR: CLEARE TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to requiring rear occupant alert systems in certain vehicles PURPOSE: To require vehicles manufactured after July 1, 2026 and sold or offered for sale in New. York as a new vehicle to be registered in this state to be equipped with a rear occupant alert system.