Bill S8737 version original - Peter Oberacker
BILL NUMBER: S8737 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the environmental conservation law, in relation to establishing a hunting mentorship program PURPOSE: Creates a hunting mentorship program to allow qualified hunters to introduce prospective hunters to the tradition of safe and responsible hunting. SUMMARY OF PROVISIONS: Section 1.