Bill S8745 version original - Roxanne J. Persaud
BILL NUMBER: S8745 SPONSOR: PERSAUD TITLE OF BILL: An act to amend the social services law, in relation to the adoption subsidies provided for disabled or hard to place children PURPOSE OR GENERAL IDEA OF BILL: Relates to the subsidies that may be provided to a parent to who adopts a disabled or hard to place child. SUMMARY OF PROVISIONS: Section 1 of the bill amends paragraph (c) of subdivision 1 of section 453 of the social services law and adds three new paragraphs (c-2), (c-3), and (c4).