Bill S8744 version original - Liz Krueger
BILL NUMBER: S8744 SPONSOR: KRUEGER TITLE OF BILL: An act to repeal section 255.17 of the penal law relating to the crime of adultery PURPOSE: To decriminalize adultery SUMMARY OF PROVISIONS: Section 1 repeals section 255.17 of the penal law. Section 2 sets the effective date. JUSTIFICATION: This outdated statute criminalizes sexual behavior between consenting adults as a Class B Misdemeanor, punishable by up to three months in prison and a fine of up to $500.