Bill S8753 version original - Anthony H. Palumbo
BILL NUMBER: S8753 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the social services law, in relation to authorizing school officials to temporarily retain custody of a suspected abused or maltreated child PURPOSE: To allow school officials to retain a child when abuse or neglect is suspected. SUMMARY OF PROVISIONS: Section 1.