Bill S8748 version original - Monica Martinez
BILL NUMBER: S8748 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the criminal procedure law, in relation to authorizing bail for principals charged with concealment of a human corpse PURPOSE: The purpose of the bill is to include the concealment of a human corpse pursuant to section 195.02 of the penal law as a bail-eligible offense. SUMMARY OF PROVISIONS: Section 1. 1 Section 1. Paragraphs (t) and (u) of subdivision 4 of section 510.