Bill S8754 version original - Anthony H. Palumbo
BILL NUMBER: S8754 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the mental hygiene law, in relation to authorizing a member of the community or a community ambassador to apply for involun- tary admission of a person alleged to be mentally ill PURPOSE: To authorize members of the community to submit an application for the involuntary admission of a person alleged to be mentally ill.. SUMMARY OF PROVISIONS: Section 1. Provides that this act shall be known as "Tony's Law". Section 2.