Bill S8750 version original - Shelley Mayer
BILL NUMBER: S8750 SPONSOR: MAYER TITLE OF BILL: An act to amend the insurance law, in relation to requiring certain insurance policies to provide coverage for neuropsychological exams for children with pediatric cancers which affect brain development or func- tion PURPOSE: To require insurers to cover comprehensive neuropsychological exams for individuals who are diagnosed with pediatric cancers or brain tumors. SUMMARY OF SPECIFIC PROVISIONS: Section 1 establishes a title for the act.