Bill S8751A version original - Anthony H. Palumbo
BILL NUMBER: S8751 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the penal law, in relation to mutilation of a human corpse PURPOSE: The purpose of this bill is to make the act of mutilating a corpse a criminal offense. SUMMARY OF PROVISIONS: Section 195.02 of the penal law is amended to include the mutilation of a human corpse as a qualifying offense.