Bill S8763 version original - Jessica Scarcella-Spanton
BILL NUMBER: S8763 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the criminal procedure law, in relation to designating certain employees of the New York city department of buildings as peace officers; and to amend the penal law, in relation to assaulting certain employees of the New York city department of buildings SUMMARY OF PROVISIONS: Section 1. Section 2.