Bill S8776A version original - Joseph P. Addabbo Jr.
BILL NUMBER: S8776 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the racing, pari-mutuel wagering and breeding law, in relation to the discretionary spending threshold for regional off-track betting corporations on food products PURPOSE OR GENERAL IDEA OF BILL: To authorize the increase the discretionary buying threshold from $15,000 to the threshold of a state agency. SUMMARY OF PROVISIONS: Section 1 of this bill amends section 507 of the racing, pari-mutuel wagering and breeding law.