Bill S7503B version A - Cordell Cleare
BILL NUMBER: S7503A SPONSOR: CLEARE TITLE OF BILL: An act to amend the general business law, in relation to requiring retailers of bicycles with electric assist and electric batteries to provide customers with an operating manual SUMMARY OF PROVISIONS: Section 1. The general business law is amended by adding a new section 391. § 2. Effective date JUSTIFICATION: To assist in the prevention of overcharging and explosion of batteries for electric assist bicycles.