Bill S6526A version A - Robert Jackson
BILL NUMBER: S6526A SPONSOR: JACKSON TITLE OF BILL: An act to amend the retirement and social security law, in relation to setting the increase to the overtime ceiling as a fixed percentage PURPOSE: The purpose of this bill is to create equity in the overtime cap between Tier 5 and Tier 6 pensions. SUMMARY OF PROVISIONS: Section 1 amends subdivision 24 of Section 501 of the Retirement and Social Security Law to equalize the amount of overtime that is pension eligible for Tiers 5 and 6.