Bill S8779 version original - James Skoufis
BILL NUMBER: S8779 SPONSOR: SKOUFIS TITLE OF BILL: An act to ratify and confirm the creation of water district no. 14 of the town of Monroe and to authorize the issuance of bonds or notes by the town of Monroe to finance certain inter-fund borrowings by water district no. 14 from such town's general fund PURPOSE OR GENERAL IDEA OF BILL: This legislation will ratify and confirm the establishment of Water District No.