Bill S5007B version B - Kristen Gonzalez
BILL NUMBER: S5007B SPONSOR: GONZALEZ TITLE OF BILL: An act to amend the state technology law, in relation to establishing the "secure our data act" SUMMARY OF PROVISIONS: Section one of this bill provides the title, which is the "secure our data act." Section two of this bill provides the legislative intent. Section three of this bill amends the State Technology Law by adding a new section 210. New subdivision 210: (1) provides definitions that will be used in this section.