Bill S8791 version original - Mark Walczyk
BILL NUMBER: S8791 SPONSOR: WALCZYK TITLE OF BILL: An act in relation to authorizing the town of Dekalb, in the county of St. Lawrence, to reduce the maximum speed limit on certain public road- ways PURPOSE: Allows the Town of Dekalb to lower the speed limit to 45 on Route 11, beginning one-quarter mile north of the Route 11 East Dekalb Road inter- section and ending one quarter mile South of the entrance to the Hermon-DeKalb Central school athletic fields, in the town of Dekalb.