Bill S8798A version original - Thomas F. O'Mara
BILL NUMBER: S8798 SPONSOR: O'MARA TITLE OF BILL: An act to amend the highway law, in relation to designating a portion of the state highway system as the "Sergeant James C. Hayes Memorial High- way" PURPOSE OR GENERAL IDEA OF BILL: To designate a portion of State Route 14 to Sergeant James C. Hayes. SUMMARY OF PROVISIONS: Section 1 adds a new section to the highway law Section 344-p. Section 344-p states the section of Route 14 that will be known as "Sergeant. James C. Hayes Memorial Highway.