Bill S8187A version A - Cordell Cleare
BILL NUMBER: S8187A SPONSOR: CLEARE TITLE OF BILL: An act to amend the state finance law and the tax law, in relation to establishing and funding the close the waitlist fund PURPOSE OF THE BILL: This bill authorizes an income tax surcharge on those earning over $10 million per annum to close the $70 million dollar waitlist gap for senior services in this state. SUMMARY OF PROVISIONS: Section 1 lays out the legislative intent for why this fund and approach are warranted.