Bill S8596A version A - Jessica Scarcella-Spanton
BILL NUMBER: S8596A SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the education law, in relation to expanding eligibility under the veterans tuition awards program PURPOSE OR GENERAL IDEA OF BILL: Relates to expanding eligibility under the veteran's tuition awards program. SUMMARY OF PROVISIONS: Section one amends paragraph D of subdivision one of section 669-a of the education law. Section two amends paragraph A of subdivision two of section 669-a of the education law.