Bill S1480A version A - Michelle Hinchey
BILL NUMBER: S1480A SPONSOR: HINCHEY TITLE OF BILL: An act to amend the executive law, in relation to establishing November twenty-sixth of each year as a day of commemoration known as "Sojourner Truth Day" PURPOSE OR GENERAL IDEA OF BILL: This legislation would establish the twenty-sixth of November as a day of commemoration known as "Sojourner Truth Day" in honor of the prolific abolitionist and women's rights advocate. A commemoration is an official day of remembrance the state recognizes.