Bill S8804 version original - James Tedisco
BILL NUMBER: S8804 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the tax law, in relation to expanding a certain tax credit for farmers to include the cost of construction housing for farm workers PURPOSE OR GENERAL IDEA OF BILL: To expand a certain tax credit for farmers to include the cost of construction of housing for farm workers. SUMMARY OF PROVISIONS: § 1.