Bill S8751A version A - Anthony H. Palumbo
BILL NUMBER: S8751A SPONSOR: PALUMBO TITLE OF BILL: An act to amend the penal law, in relation to concealment and/or mutila- tion of a human corpse; and to amend the criminal procedure law, in relation to authorizing bail for principals charged with concealment and/or mutilation of a human corpse PURPOSE: To make the act of mutilating and/or concealing a human corpse a crimi- nal offense and bail eligible. SUMMARY OF PROVISIONS: Section 1. Section 195.