Bill S8806 version original - James Skoufis
BILL NUMBER: S8806 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the town law and the public officers law, in relation to waiving the residency requirement for the comptroller and building inspector in the town of Chester, county of Orange PURPOSE: To waive the residency requirement for the comptroller and building inspector in the town of Chester, county of Orange. SUMMARY OF PROVISIONS: Section 1: Amends section 23 of the Chester town law by adding subdivi- sion 36.